requirements.txt
Flask==3.0.0
flask-socketio==5.3.6
gunicorn==21.2.0
PyJWT==2.8.0
eventlet==0.33.3
